API change to expose more `game` attributes
