+++ .vscode/settings.json
@@ -2,6 +2,8 @@
     "cSpell.words": [
         "fabius",
         "quintus",
-        "valerius"
+        "rorapp",
+        "valerius",
+        "viewsets"
     ]
 }
