+++ backend/rorapp/serializers/__init__.py
@@ -1,4 +1,3 @@
 from .game import GameSerializer
 from .token import TokenObtainPairByEmailViewSerializer
-from .user import UserSerializer
-from .user_detail import UserDetailSerializer
+from .user import UserSerializer, UserDetailSerializer
