--- backend/rorapp/serializers/user_detail.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from django.contrib.auth.models import User
-
-
-class UserDetailSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('username', 'email')
