+++ backend/rorapp/views/__init__.py
@@ -1,4 +1,3 @@
 from .game import GameViewSet
 from .token import TokenObtainPairByEmailView
-from .user import UserViewSet
-from .user_detail import UserDetailViewSet
+from .user import UserViewSet, UserDetailViewSet
