+++ backend/rorapp/views/game.py
@@ -9,6 +9,12 @@ class GameViewSet(viewsets.ModelViewSet):
     Retrieve all games.
     """
 
-    queryset = Game.objects.all().order_by('name')
+    queryset = Game.objects.all().order_by('-creation_date')
     serializer_class = GameSerializer
     permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        
+        # Prefetch the owner's username to eliminate the need to make additional queries to the database
+        return queryset.prefetch_related('owner')
