+++ backend/rorapp/views/user.py
@@ -1,6 +1,6 @@
 from rest_framework import viewsets
 from django.contrib.auth.models import User
-from rorapp.serializers import UserSerializer
+from rorapp.serializers import UserSerializer, UserDetailSerializer
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 
@@ -15,4 +15,17 @@ class UserViewSet(viewsets.ViewSet):
     def list(self, request):
         queryset = User.objects.all()
         serializer = UserSerializer(queryset, many=True)
+        return Response(serializer.data)
+
+
+class UserDetailViewSet(viewsets.ViewSet):
+    """
+    Retrieve the username and email of the current user.
+    """
+
+    permission_classes = [IsAuthenticated]
+
+    def list(self, request):
+        user = User.objects.get(username=request.user.username)
+        serializer = UserDetailSerializer(user)
         return Response(serializer.data)
